refactor: extract invite procedures from organization router into dedicated invite router
Move 6 invite-related procedures (createInvites, deleteInvite,
getOrganizationPendingInvites, createInviteRequest, approveInvite,
acceptInvite) into their own router for separation of concerns.
Organization router drops from ~1434 to ~660 lines.

Closes #2877
