+++ agentic-e2e-tests/tests/members/steps.ts
@@ -268,7 +268,7 @@ export async function seedWaitingApprovalInvite({
   teamId: string;
 }) {
   const response = await page.request.post(
-    "/api/trpc/organization.createInviteRequest",
+    "/api/trpc/invite.createInviteRequest",
     {
       data: {
         json: {
