+++ langwatch/src/pages/settings/members.tsx
@@ -117,7 +117,7 @@ function MembersList({
   } = useDisclosure();
 
   const pendingInvites =
-    api.organization.getOrganizationPendingInvites.useQuery(
+    api.invite.getOrganizationPendingInvites.useQuery(
       {
         organizationId: organization?.id ?? "",
       },
