+++ langwatch/src/server/api/routers/__tests__/invite.integration.test.ts
@@ -277,7 +277,7 @@ describe("Organization Invites Integration", () => {
       /** @scenario "Member cannot request invitation with ADMIN role" */
       it("fails with validation error", async () => {
         await expect(
-          memberCaller.organization.createInviteRequest({
+          memberCaller.invite.createInviteRequest({
             organizationId,
             invites: [
               {
@@ -294,7 +294,7 @@ describe("Organization Invites Integration", () => {
 
     describe("when member requests invitation with MEMBER role", () => {
       it("creates invitation with WAITING_APPROVAL status", async () => {
-        const results = await memberCaller.organization.createInviteRequest({
+        const results = await memberCaller.invite.createInviteRequest({
           organizationId,
           invites: [
             { email: "user@example.com", role: "MEMBER", teamIds: teamId },
@@ -306,7 +306,7 @@ describe("Organization Invites Integration", () => {
 
       /** @scenario "Member request sets requestedBy to the requesting user" */
       it("sets requestedBy to the requesting user's ID", async () => {
-        const results = await memberCaller.organization.createInviteRequest({
+        const results = await memberCaller.invite.createInviteRequest({
           organizationId,
           invites: [
             { email: "user@example.com", role: "MEMBER", teamIds: teamId },
@@ -318,7 +318,7 @@ describe("Organization Invites Integration", () => {
 
       /** @scenario "Invitation request has no expiration while awaiting approval" */
       it("creates invitation with null expiration", async () => {
-        const results = await memberCaller.organization.createInviteRequest({
+        const results = await memberCaller.invite.createInviteRequest({
           organizationId,
           invites: [
             { email: "user@example.com", role: "MEMBER", teamIds: teamId },
@@ -330,7 +330,7 @@ describe("Organization Invites Integration", () => {
 
       /** @scenario "No email is sent when a member creates an invitation request" */
       it("does not send invitation email", async () => {
-        await memberCaller.organization.createInviteRequest({
+        await memberCaller.invite.createInviteRequest({
           organizationId,
           invites: [
             { email: "user@example.com", role: "MEMBER", teamIds: teamId },
@@ -343,7 +343,7 @@ describe("Organization Invites Integration", () => {
 
     describe("when member requests multiple invitations at once", () => {
       it("creates a WAITING_APPROVAL record for each email", async () => {
-        const results = await memberCaller.organization.createInviteRequest({
+        const results = await memberCaller.invite.createInviteRequest({
           organizationId,
           invites: [
             { email: "multi-a@example.com", role: "MEMBER", teamIds: teamId },
@@ -363,7 +363,7 @@ describe("Organization Invites Integration", () => {
 
       it("rejects duplicate emails in a single payload and creates no invites", async () => {
         await expect(
-          memberCaller.organization.createInviteRequest({
+          memberCaller.invite.createInviteRequest({
             organizationId,
             invites: [
               { email: "dupe@example.com", role: "MEMBER", teamIds: teamId },
@@ -391,7 +391,7 @@ describe("Organization Invites Integration", () => {
       /** @scenario "Duplicate detection across PENDING and WAITING_APPROVAL statuses" */
       it("fails with duplicate invitation error", async () => {
         // Create initial WAITING_APPROVAL invite
-        await memberCaller.organization.createInviteRequest({
+        await memberCaller.invite.createInviteRequest({
           organizationId,
           invites: [
             { email: "existing@example.com", role: "MEMBER", teamIds: teamId },
@@ -400,7 +400,7 @@ describe("Organization Invites Integration", () => {
 
         // Try to create another for the same email
         await expect(
-          memberCaller.organization.createInviteRequest({
+          memberCaller.invite.createInviteRequest({
             organizationId,
             invites: [
               {
@@ -428,14 +428,14 @@ describe("Organization Invites Integration", () => {
       it("transitions status to PENDING", async () => {
         // Create WAITING_APPROVAL invite
         const results =
-          await memberCaller.organization.createInviteRequest({
+          await memberCaller.invite.createInviteRequest({
             organizationId,
             invites: [
               { email: "user@example.com", role: "MEMBER", teamIds: teamId },
             ],
           });
 
-        const result = await adminCaller.organization.approveInvite({
+        const result = await adminCaller.invite.approveInvite({
           inviteId: results[0]!.invite.id,
           organizationId,
         });
@@ -447,14 +447,14 @@ describe("Organization Invites Integration", () => {
         const beforeApproval = Date.now();
 
         const results =
-          await memberCaller.organization.createInviteRequest({
+          await memberCaller.invite.createInviteRequest({
             organizationId,
             invites: [
               { email: "user@example.com", role: "MEMBER", teamIds: teamId },
             ],
           });
 
-        const result = await adminCaller.organization.approveInvite({
+        const result = await adminCaller.invite.approveInvite({
           inviteId: results[0]!.invite.id,
           organizationId,
         });
@@ -470,14 +470,14 @@ describe("Organization Invites Integration", () => {
       /** @scenario "Email is sent when admin approves an invitation request" */
       it("sends invitation email", async () => {
         const results =
-          await memberCaller.organization.createInviteRequest({
+          await memberCaller.invite.createInviteRequest({
             organizationId,
             invites: [
               { email: "user@example.com", role: "MEMBER", teamIds: teamId },
             ],
           });
 
-        await adminCaller.organization.approveInvite({
+        await adminCaller.invite.approveInvite({
           inviteId: results[0]!.invite.id,
           organizationId,
         });
@@ -508,7 +508,7 @@ describe("Organization Invites Integration", () => {
         });
 
         await expect(
-          memberCaller.organization.approveInvite({
+          memberCaller.invite.approveInvite({
             inviteId: invite.id,
             organizationId,
           })
@@ -524,7 +524,7 @@ describe("Organization Invites Integration", () => {
         mockGetActivePlan.mockResolvedValue(makeTestPlan({ maxMembers: 10 }));
 
         const results =
-          await memberCaller.organization.createInviteRequest({
+          await memberCaller.invite.createInviteRequest({
             organizationId,
             invites: [
               {
@@ -542,7 +542,7 @@ describe("Organization Invites Integration", () => {
         mockGetActivePlan.mockResolvedValue(makeTestPlan({ maxMembers: 2 }));
 
         await expect(
-          adminCaller.organization.approveInvite({
+          adminCaller.invite.approveInvite({
             inviteId,
             organizationId,
           })
@@ -582,7 +582,7 @@ describe("Organization Invites Integration", () => {
           },
         });
 
-        await adminCaller.organization.deleteInvite({
+        await adminCaller.invite.deleteInvite({
           inviteId: invite.id,
           organizationId,
         });
@@ -631,7 +631,7 @@ describe("Organization Invites Integration", () => {
         });
 
         const invites =
-          await adminCaller.organization.getOrganizationPendingInvites({
+          await adminCaller.invite.getOrganizationPendingInvites({
             organizationId,
           });
 
@@ -655,7 +655,7 @@ describe("Organization Invites Integration", () => {
         });
 
         const invites =
-          await adminCaller.organization.getOrganizationPendingInvites({
+          await adminCaller.invite.getOrganizationPendingInvites({
             organizationId,
           });
 
@@ -683,7 +683,7 @@ describe("Organization Invites Integration", () => {
           callOrder.push("email-sent");
         });
 
-        const results = await adminCaller.organization.createInvites({
+        const results = await adminCaller.invite.createInvites({
           organizationId,
           invites: [
             { email: "batch-a@example.com", role: "MEMBER", teamIds: teamId },
@@ -711,7 +711,7 @@ describe("Organization Invites Integration", () => {
           .mockResolvedValueOnce(undefined) // first email succeeds
           .mockRejectedValueOnce(new Error("SMTP failure")); // second email fails
 
-        const results = await adminCaller.organization.createInvites({
+        const results = await adminCaller.invite.createInvites({
           organizationId,
           invites: [
             { email: "ok-email@example.com", role: "MEMBER", teamIds: teamId },
@@ -761,7 +761,7 @@ describe("Organization Invites Integration", () => {
       it("still approves the invitation", async () => {
         // Create WAITING_APPROVAL invite
         const results =
-          await memberCaller.organization.createInviteRequest({
+          await memberCaller.invite.createInviteRequest({
             organizationId,
             invites: [
               { email: "user@example.com", role: "MEMBER", teamIds: teamId },
@@ -773,7 +773,7 @@ describe("Organization Invites Integration", () => {
           new Error("Email service unavailable")
         );
 
-        const result = await adminCaller.organization.approveInvite({
+        const result = await adminCaller.invite.approveInvite({
           inviteId: results[0]!.invite.id,
           organizationId,
         });
@@ -790,7 +790,7 @@ describe("Organization Invites Integration", () => {
 
       it("returns emailNotSent as fallback indicator", async () => {
         const results =
-          await memberCaller.organization.createInviteRequest({
+          await memberCaller.invite.createInviteRequest({
             organizationId,
             invites: [
               { email: "user@example.com", role: "MEMBER", teamIds: teamId },
@@ -801,7 +801,7 @@ describe("Organization Invites Integration", () => {
           new Error("Email service unavailable")
         );
 
-        const result = await adminCaller.organization.approveInvite({
+        const result = await adminCaller.invite.approveInvite({
           inviteId: results[0]!.invite.id,
           organizationId,
         });
@@ -822,7 +822,7 @@ describe("Organization Invites Integration", () => {
         mockGetActivePlan.mockResolvedValue(makeTestPlan({ maxMembers: 10 }));
 
         const results =
-          await memberCaller.organization.createInviteRequest({
+          await memberCaller.invite.createInviteRequest({
             organizationId,
             invites: [
               {
@@ -843,7 +843,7 @@ describe("Organization Invites Integration", () => {
 
         // Step 3: Admin attempts to approve the invite
         await expect(
-          adminCaller.organization.approveInvite({
+          adminCaller.invite.approveInvite({
             inviteId,
             organizationId,
           })
@@ -876,7 +876,7 @@ describe("Organization Invites Integration", () => {
         });
 
         await expect(
-          memberCaller.organization.createInviteRequest({
+          memberCaller.invite.createInviteRequest({
             organizationId,
             invites: [
               { email: "new@example.com", role: "MEMBER", teamIds: teamId },
